requirements.txt
jsonlines==4.0.0
openai==1.34.0
python-dotenv==1.0.1
pymupdf==1.24.5